[WEB] refactor : 컴포넌트 폴더로 정리
레이블 페이지 컴포넌트들을 레이블 폴더와 공용으로 쓸 수 있을만한 컴포넌트를 폴더로 정리했습니다.
