+++ frontend/src/components/label/LabelCount.js
