+++ frontend/src/templates/LabelList.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
-import ListHead from '@components/ListHead';
-import LabelCount from '@components/LabelCount';
-import List from '@components/List';
+import ListHead from '@components/label/ListHead';
+import LabelCount from '@components/label/LabelCount';
+import List from '@components/public/List';
 import { useLabelState } from '@contexts/label';
 
 const SummaryDiv = styled.div`
